requirements.txt
Flask>=2.3.0,<3.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-Login>=0.6.0,<1.0.0
Flask-WTF>=1.1.0,<2.0.0
requests>=2.28.0,<3.0.0
beautifulsoup4>=4.11.0,<5.0.0
